🚨 [Tests]: api/dataテストファイルを機能別に分割
component.test.ts を以下に分割:
- interface.test.ts: インターフェーステスト
- from-response.test.ts: レスポンス変換テスト
- fetch-all.test.ts: API取得テスト
- group-by-page.test.ts: グループ化テスト
- filter-by-name.test.ts: フィルタリングテスト

file.test.ts を以下に分割:
- interface.test.ts: インターフェーステスト
- from-response.test.ts: レスポンス変換テスト
- fetch.test.ts: API取得テスト
- fetch-nodes.test.ts: ノード取得テスト
- utility-functions.test.ts: ユーティリティ関数テスト

image-export.test.ts を以下に分割:
- from-options.test.ts: オプション作成テスト
- fetch.test.ts: API取得テスト
- utility-functions.test.ts: ユーティリティ関数テスト
- validate.test.ts: バリデーションテスト
- fetch-in-batches.test.ts: バッチ処理テスト

image.test.ts を以下に分割:
- interface.test.ts: インターフェーステスト
- from-options.test.ts: オプション作成テスト
- fetch.test.ts: API取得テスト
- fetch-batch.test.ts: バッチ処理テスト

テスト構造のベストプラクティスに従い、機能ごとにファイルを分割して管理性を向上
